requirements.txt
pluxee-api
requests